% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{find_stata}
\alias{find_stata}
\title{Find and set Stata binary path}
\usage{
find_stata()
}
\description{
Set Stata binary (among found alternatives) path. These settings are
lost when R is closed, therefore you should consider adding a
\code{options("RStata.StataPath")} line in your \code{.Rprofile}.
}
\examples{
\dontrun{
find_stata()
}
}
