% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr.R
\name{set_stata_engine}
\alias{set_stata_engine}
\title{Knitr engine for Stata}
\usage{
set_stata_engine(
  bin = getOption("RStata.StataPath",
    stop("You need to set up a Stata path; ?find_stata"))
)
}
\arguments{
\item{bin}{Path to Stata bin directory, if \code{NULL} then statarepl
will set it to \code{getOption("RStata.StataPath", stop("You need to set up a Stata path; ?find_stata"))}, if you don't have \code{options("RStata.StataPath")} setted, you can search and set it use \code{\link{find_stata}}}
}
\description{
Knitr engine for Stata
}
\examples{
\dontrun{
statarepl::set_stata_engine()
}
}
