% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{stata_repl}
\alias{stata_repl}
\title{Stata REPL}
\usage{
stata_repl(
  bin = getOption("RStata.StataPath",
    stop("You need to set up a Stata path; ?find_stata"))
)
}
\arguments{
\item{bin}{Path to Stata bin directory, if \code{NULL} then statarepl
will set it to \code{getOption("RStata.StataPath", stop("You need to set up a Stata path; ?find_stata"))}, if you don't have \code{options("RStata.StataPath")} setted, you can search and set it use \code{\link{find_stata}}}
}
\value{
a Stata REPL
}
\description{
Stata REPL
}
\examples{
\dontrun{
stata_repl()
}
}
